Symbol look is working now.
graps now manages source context for each loaded grammar (in editor) or which is referenced in such a grammar (tokenVocab or import). It takes care not to load a grammar more than once and freeing the source contextx by keeping a reference count. Symbol override in imported grammars works like ANTLR is handling it.