Lots of fine tuning in this patch.
Now there is a big test file for most of the settings, which are used in various combinations.
Another test file (not yet used) will deal with various alignments and their combined use. It will at the same time also be used for performance testing (almost 10K lines to format).